Extract trace classification logic into helper file
